MA-128 refactor operaor endpoint ui
