+++ component/ui/src/app/commons/components/common-data-table/common-data-table.component.scss
