+++ component/ui/src/app/commons/models/operator-onboarding-data-models.ts
