+++ component/ui/src/app/config/api.endpoints.ts
