+++ component/ui/src/app/data-providers/operator-onboarding-data.service.ts
