+++ component/ui/src/app/operatoronboarding/add/add-main/operatoronboarding.component.scss
