+++ component/ui/src/app/operatoronboarding/add/add-token/token.component.ts
