+++ component/ui/src/app/operatoronboarding/create-operator-endpoint/create-operator-endpoint.component.html
