+++ component/ui/src/app/operatoronboarding/operatoronboarding.routes.ts
