+++ component/ui/src/app/operatoronboarding/view-operators/view-operators.component.scss
