+++ component/ui/src/app/operatoronboarding/view-operators/view-operators.component.spec.ts
