+++ component/ui/src/assets/json/get.operators.json
